src/ai_workflow/pyproject.toml
[tool.poetry]
name = "ai-workflow"
version = "0.1.0"
description = ""
authors = ["Ayo Tosan Samuel <46663234+Stosan@users.noreply.github.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.69.0"
grpcio-tools = "^1.69.0"
grpc-interceptor = "^0.15.4"
python-dotenv = "^1.0.1"
langgraph = "^0.2.52"
langchain-community = "^0.3.7"
psycopg2-binary = "^2.9.10"
sqlalchemy = "2.0.35"
vertexai = "^1.71.1"
selenium = "^4.26.1"
pillow = "^10.2.0"
beautifulsoup4 = "^4.12.3"
celery = {extras = ["redis"], version = "^5.4.0"}
pymupdf = "^1.24.14"
colorama = "^0.4.6"
jinja2 = "^3.1.4"
pika = "^1.3.2"
langchain-openai = "^0.2.10"
gunicorn = "^23.0.0"
langchain-google-community = {extras = ["featurestore"], version = "^2.0.3"}
google-cloud-storage = "^2.18.2"
google-cloud-bigquery = "^3.27.0"
pypdf = "^5.1.0"
llama-index-multi-modal-llms-gemini = "^0.4.1"
llama-index-llms-gemini = "^0.4.1"
llama-index-llms-vertex = "^0.4.2"
pytest = "^8.3.4"
playwright = "^1.49.1"
pydantic = ">=2.0.0,<3.0.0"
pydantic-settings = ">=2.0.0"
langchain = ">=0.0.350"
langchain-google-vertexai = ">=0.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false